pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openmetadata-semantic-onboarding"
version = "0.1.0"
description = "Reusable OpenMetadata-backed semantic onboarding framework for local databases."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "alembic>=1.16,<2",
  "fastapi>=0.116,<1",
  "httpx>=0.28,<1",
  "jinja2>=3.1,<4",
  "orjson>=3.10,<4",
  "pydantic>=2.11,<3",
  "pydantic-settings>=2.10,<3",
  "PyYAML>=6.0,<7",
  "python-dotenv>=1.1,<2",
  "python-multipart>=0.0.20,<1",
  "pymysql>=1.1,<2",
  "psycopg[binary]>=3.2,<4",
  "sqlalchemy>=2.0,<3",
  "sqlglot>=27,<28",
  "tabulate>=0.9,<1",
  "typer>=0.16,<1",
  "uvicorn[standard]>=0.35,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
  "pytest-asyncio>=1.2,<2",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
